Do not spit out HTML in a non-cli context.
Spoon's exception handler might be nice when viewing the error message
in a web context, ie. with a browser, but when running stuff from the
command line, this is not readable, further more, it throws a notice
because $_SERVER['REQUEST_METHOD'] is not known in php-cli.

By checking in which context we are before registering the exception
handler, we prevent this.  In addition the cli check is put in the main
Spoon class since it may be useful in other places too.