Type check elements of multi-checkbox.
The multi checkbox works somewhat different than the drop down menu,
where you pass a mapping value => label.  This has lead to confusion
when working with the multi checkbox whereby checkboxes are displayed,
but only the first character of the label is shown.

At first I thought the isset($value['label']/['value']) was to blame,
but, while I'm still inclined to think using isset() for array key
existance checks is not a good idea, this was not the root of the
problem.  The fact that code in user land passes an assoc. array instead
of an array of assoc. arrays, is however.

By checking the type of each value, prevents this, while at the same
time leaving the rest of the production code in tact; none of the
isset() checks needed changing.  This check is also essential: using
isset($value['label]) doesn't throw an exception at all if $value was a
string instead of an assoc. array and, on the other hand: using
array_key_exists('label', $value) would throw a not so clear error
complaining about the second argument to array_key_exists() not being an
array.