Spoon::dump: improve Xdebug var_dump detection
It is possible to have Xdebug loaded, but not have it override the var_dump
output. This leaves the output lacking.

Note that Xdebug is smart about the SAPI: it only outputs HTML when it is
expected to. The CLI does not get HTML. Spoon::dump() would profit from a
similar approach.